test(server): updated methods test file and refactored
